+++ internal/authz/bitbucketserver/testdata/vcr/BitbucketServer.yaml
