+++ internal/authz/bitbucketserver/testdata/vcr/Validate-no-problems-when-authenticated-as-admin.yaml
