+++ internal/authz/bitbucketserver/testdata/vcr/Validate-problems-when-authenticated-as-non-admin.yaml
